fix(build): resolve ambiguous value::number(long long) overload under C++17
C++17 stricter overload resolution makes long long int ambiguous between
int64_t and other overloads. Cast to int64_t explicitly.

+++ src/trade_client.cpp
@@ -1054,7 +1054,7 @@ namespace TIGER_API {
         value obj = value::object(true);
         obj[P_ACCOUNT] = get_account_param(account);
         set_secret_key(obj);
-        obj[U("id")] = value::number(exercise_id);
+        obj[U("id")] = value::number(static_cast<int64_t>(exercise_id));
         return post(OPTION_EXERCISE_CANCEL, obj);
     }
